Jail implementation (#37)
* jail: initial implementation

Key Terms:

Jail - rule that informs Guardian to ban client IP addresses that surpass a configured limit. They are banned for the entire BanDuration from making any requests.

Prisoner - a client IP that is currently banned due to any of the configured Jails. Also referred to as a client IP that has been jailed.

Jailer - uses the jails and a counter to keep track of which Client IPs should be prisoners.

PrisonerStore - used by the Jailer to get the current prisoners and also add prisoners as needed

Intended Usage
While we can use Global Rate Limits and Route Rate Limits to provide rate limits, it can be helpful to block client ips for longer periods of time than the Limit.Duration.

Previously, we could only do this by manually adding users to the blacklist.

With Jails, we can have a powerful tool for automatically block malicious IPs from using any services behind Guardian, for a configurable amount of time.